Forget about system libs
First of all, since there are already inbuilt libs the system
libs won't be used anyways unless I swap the order of it in
the library path.

But using system libs does not really make sense. They could be
broken. And not using inbuilt libs and copying system libs to
rootdir makes even less sense because they could be either broken
or not exist at all.

So I decided to rather import working libs so that we don't
have to worry about the system at all.